It's alive (again)!
I have gotten to the point where everything is more or less (more more
than less) encapsulated inside of the LitOptions class. It might make
sense to refactor this further at some point, in fact, it's already
becoming a bit unweildy, so I feel that it will have to be, but I am
satisfied with the level of abstraction I have acheived at this venture.

There is certainly work to be done, but I feel that this skeleton is the
best place to build from. It's still pre-alpha, but much less pre-alpha
then before.